created foodsTable object
