+++ lib/foods_index.js
@@ -1,6 +1,7 @@
 var jQuery = require('jquery');
 window.$ = jQuery;
-var food = require('./food.js')
+var food = require('./food.js');
+var foodsTable = require('./foods_table.js');
 
 function populateFoods(){
   var currentFoods = localStorage.getItem('foods')
@@ -8,14 +9,9 @@ function populateFoods(){
     var currentFoodsJSON = JSON.parse(currentFoods);
     for (var i = 0; i < currentFoodsJSON.length; i++) {
       var foodItem = new food(currentFoodsJSON[i]['name'], currentFoodsJSON[i]['calories']);
-      appendToFoods(foodItem);
+      foodsTable.appendTo(foodItem);
     }
   }
-}
-
-function appendToFoods(foodItem){
-  var deleteButton = '<button class="delete-button"><b>-</b></button>'
-  $('#foods-table tr:first').after('<tr><td class="food-cell food-name-cell">' + foodItem.name + '</td><td class="food-cell food-calorie-cell">' + foodItem.calories + '</td><td class="delete-cell">' + deleteButton + '</td></tr>')
 }
 
 function clearFoodForm(){
@@ -47,35 +43,6 @@ function removeErrors(){
   $('#calorie-error').addClass('hidden');
 }
 
-function filterFoods(){
-  var input = document.getElementById('food-filter')
-  var filter = input.value.toUpperCase();
-  var table = document.getElementById('foods-table');
-  var tr = table.getElementsByTagName('tr');
-
-  for (var i = 0; i < tr.length; i++) {
-    td = tr[i].getElementsByTagName("td")[0];
-    if (td) {
-      if (td.innerHTML.toUpperCase().indexOf(filter) > -1) {
-        tr[i].style.display = "";
-      } else {
-        tr[i].style.display = "none";
-      }
-    }
-  }
-}
-
-function updateCell(cell, originalContent, newContent){
-  if(cell.hasClass('food-name-cell')){
-    var foodItem = food.find(originalContent);
-    foodItem.update('name', newContent);
-  } else if(cell.hasClass('food-calorie-cell')){
-    var foodName = cell.siblings('.food-name-cell').html();
-    var foodItem = food.find(foodName);
-    foodItem.update('calories', newContent)
-  }
-}
-
 $(document).ready(function(){
   populateFoods();
 
@@ -91,7 +58,7 @@ $(document).ready(function(){
     } else {
       foodItem = new food(name, calories)
       foodItem.store();
-      appendToFoods(foodItem);
+      foodsTable.appendTo(foodItem);
       clearFoodForm();
       removeErrors();
     }
@@ -104,7 +71,7 @@ $(document).ready(function(){
   });
 
   $('#food-filter').keyup(function(){
-    filterFoods();
+    foodsTable.filter();
   })
 
   $('#foods-table').on('click', '.food-cell', function(e){
@@ -123,7 +90,7 @@ $(document).ready(function(){
       if(e.which == 13 || e.type == 'blur') {
         var newContent = $(this).val();
         $(this).parent().text(newContent);
-        updateCell(cell, originalContent, newContent);
+        foodsTable.updateCell(cell, originalContent, newContent);
       }
     });
   });
